requirements.txt
moviepy>=1.0.3
Pillow>=9.0.0
numpy>=1.21.0
pyyaml>=6.0
flask>=3.0.0
